Only try to rmdir() the tmp directory if we have one.
2004-01-27  Christian Persch  <chpe@cvs.gnome.org>

	* lib/ephy-file-helpers.c: (ephy_file_helpers_shutdown):

	Only try to rmdir() the tmp directory if we have one.

+++ lib/ephy-file-helpers.c
@@ -188,9 +188,12 @@ ephy_file_helpers_shutdown (void)
 	g_list_free (del_on_exit);
 	del_on_exit = NULL;
 
-	rmdir (tmp_dir);
-	g_free (tmp_dir);
-	tmp_dir = NULL;
+	if (tmp_dir != NULL)
+	{
+		rmdir (tmp_dir);
+		g_free (tmp_dir);
+		tmp_dir = NULL;
+	}
 
 	g_free (dot_dir);
 	dot_dir = NULL;
